Update solution for metrics-f1-micro

+++ metrics-f1-micro/metrics-f1-micro.py
@@ -2,8 +2,8 @@ def f1_micro(y_true, y_pred) -> float:
     """
     Compute micro-averaged F1 for multi-class integer labels.
     """
-    y_true = np.array(y_true, dtype=np.float64)
-    y_pred = np.array(y_pred, dtype=np.float64)
+    y_true = np.array(y_true, dtype=np.int64)
+    y_pred = np.array(y_pred, dtype=np.int64)
 
     if y_true.shape[0] != y_pred.shape[0]:
         return None
